ssa: remove Recover entry point
The entry point for Recover isn't very useful and complicates some
logic due to adding a second entry to a function. Furthermore, it
doesn't really make sense as an entry point to the function. The work
it is doing is better described by panic jumping to the exit node,
and possibly extra logic outside the function.