ssa: encode memory dependencies
Make explicit the fact that certain instructions modify memory state,
by accepting and returning memory states. This way we can establish a
dependency on the order of instructions, e.g. that two function calls
must not be reordered.

Our model of memory is very simple: there is a single, global memory
state. Any instruction that reads from memory accepts the memory state
as an input, and any instruction that writes to memory returns a new
memory state. The implication is that our model cannot deduce that two
stores two two different memory locations are independent of each
other.

The main envisioned use of the memory state is to simplify the future
implementation of CSE. Because the memory state is now an operand of
certain instructions, instructions that are identical aside from the
memory state won't be eliminated.

At a later point we may analyze function bodies to determine functions
that are pure, to encode the fact that they do not modify memory.