ssa: remove notion of memory state
While encoding dependencies on and modifications to the memory state
would allow for a simpler implementation of CSE, it almost universally
makes other analyses more difficult to implement and comprehend. The
introduction of memory state moves the SSA representation further from
the source representation, and adds extra layers of indirection
between most operations and the results that analyses actually care
about.

Since tracking memory in SSA directly is in no way required for
implementing CSE, and no other analysis benefits from it, remove it.